🚚Modify : 네이밍 변경, 라우터 최신화
